requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML and Computer Vision
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0
transformers==4.36.2
Pillow==10.2.0

# Additional ML dependencies for BLIP-2
accelerate==0.26.1
sentencepiece==0.1.99

# Utilities
pydantic==2.5.3
numpy==1.26.3
opencv-python==4.9.0.80
